Use Mir to calculate priority, test for priority
