+++ logos-derive/src/graph/regex.rs
@@ -1,122 +1,16 @@
 use std::fmt::Debug;
 use std::cmp::min;
-use std::convert::TryFrom;
 
-use regex_syntax::hir::{Class, ClassUnicode, Hir, HirKind, Literal, RepetitionKind};
-use regex_syntax::ParserBuilder;
 use utf8_ranges::Utf8Sequences;
 
 use crate::graph::{Graph, Disambiguate, Node, NodeId, ReservedId, Range, Rope, Fork};
-use crate::error::{Error, Result};
-
-/// Middle Intermediate Representation of the regex, built from
-/// `regex_syntax`'s `Hir`. The goal here is to strip and canonicalize
-/// the tree, so that we don't have to do transformations later on the
-/// graph, with the potential of running into looping references.
-#[derive(Clone, Debug)]
-enum Mir {
-    Empty,
-    Loop(Box<Mir>),
-    Maybe(Box<Mir>),
-    Concat(Vec<Mir>),
-    Alternation(Vec<Mir>),
-    Class(Class),
-    Literal(Literal),
-}
-
-impl TryFrom<Hir> for Mir {
-    type Error = Error;
-
-    fn try_from(hir: Hir) -> Result<Mir> {
-        match hir.into_kind() {
-            HirKind::Empty => {
-                Ok(Mir::Empty)
-            },
-            HirKind::Concat(concat) => {
-                let mut out = Vec::with_capacity(concat.len());
-
-                fn extend(mir: Mir, out: &mut Vec<Mir>) {
-                    match mir {
-                        Mir::Concat(nested) => {
-                            for child in nested {
-                                extend(child, out);
-                            }
-                        },
-                        mir => out.push(mir),
-                    }
-                }
-
-                for hir in concat {
-                    extend(Mir::try_from(hir)?, &mut out);
-                }
-
-                Ok(Mir::Concat(out))
-            },
-            HirKind::Alternation(alternation) => {
-                let alternation = alternation
-                    .into_iter()
-                    .map(Mir::try_from)
-                    .collect::<Result<_>>()?;
-
-                Ok(Mir::Alternation(alternation))
-            },
-            HirKind::Literal(literal) => {
-                Ok(Mir::Literal(literal))
-            },
-            HirKind::Class(class) => {
-                Ok(Mir::Class(class))
-            },
-            HirKind::Repetition(repetition) => {
-                if !repetition.greedy {
-                    Err("#[regex]: non-greedy parsing is currently unsupported.")?;
-                }
-
-                let kind = repetition.kind;
-                let mir = Mir::try_from(*repetition.hir)?;
-
-                match kind {
-                    RepetitionKind::ZeroOrOne => {
-                        Ok(Mir::Maybe(Box::new(mir)))
-                    },
-                    RepetitionKind::ZeroOrMore => {
-                        Ok(Mir::Loop(Box::new(mir)))
-                    },
-                    RepetitionKind::OneOrMore => {
-                        Ok(Mir::Concat(vec![
-                            mir.clone(),
-                            Mir::Loop(Box::new(mir)),
-                        ]))
-                    },
-                    RepetitionKind::Range(..) => {
-                        Err("#[regex]: {n,m} repetition range is currently unsupported.")?
-                    },
-                }
-            },
-            HirKind::Group(group) => {
-                Mir::try_from(*group.hir)
-            },
-            HirKind::WordBoundary(_) => {
-                Err("#[regex]: word boundaries are currently unsupported.")?
-            },
-            HirKind::Anchor(_) => {
-                Err("#[regex]: anchors in #[regex] are currently unsupported.")?
-            },
-        }
-    }
-}
+use crate::mir::{Mir, Literal, Class, ClassUnicode};
 
 impl<Leaf: Disambiguate + Debug> Graph<Leaf> {
-    pub fn regex(&mut self, utf8: bool, source: &str, then: NodeId) -> Result<(usize, NodeId)> {
-        let mut builder = ParserBuilder::new();
+    pub fn regex(&mut self, mir: Mir, then: NodeId) -> NodeId {
+        let (_ ,id) = self.parse_mir(mir, then, None, None);
 
-        if !utf8 {
-            builder.allow_invalid_utf8(true).unicode(false);
-        }
-
-        let hir = builder.build().parse(source)?;
-        let mir = Mir::try_from(hir.clone())?;
-
-        Ok(self.parse_mir(mir, then, None, None))
+        id
     }
 
     fn parse_mir(
@@ -348,12 +242,15 @@ mod tests {
     fn rope() {
         let mut graph = Graph::new();
 
-        let leaf = graph.push(Node::Leaf("LEAF"));
-        let (len, parsed) = graph.regex(true, "foobar", leaf).unwrap();
+        let mir = Mir::utf8("foobar").unwrap();
 
-        assert_eq!(len, 12);
+        assert_eq!(mir.priority(), 12);
+
+        let leaf = graph.push(Node::Leaf("LEAF"));
+        let id = graph.regex(mir, leaf);
+
         assert_eq!(
-            graph[parsed],
+            graph[id],
             Node::Rope(Rope::new("foobar", leaf)),
         )
     }
@@ -362,12 +259,15 @@ mod tests {
     fn alternation() {
         let mut graph = Graph::new();
 
-        let leaf = graph.push(Node::Leaf("LEAF"));
-        let (len, parsed) = graph.regex(true, "a|b", leaf).unwrap();
+        let mir = Mir::utf8("a|b").unwrap();
 
-        assert_eq!(len, 2);
+        assert_eq!(mir.priority(), 2);
+
+        let leaf = graph.push(Node::Leaf("LEAF"));
+        let id = graph.regex(mir, leaf);
+
         assert_eq!(
-            graph[parsed],
+            graph[id],
             Node::Fork(
                 Fork::new()
                     .branch(b'a', leaf)
@@ -380,15 +280,18 @@ mod tests {
     fn repeat() {
         let mut graph = Graph::new();
 
-        let leaf = graph.push(Node::Leaf("LEAF"));
-        let (len, parsed) = graph.regex(true, "[a-z]*", leaf).unwrap();
+        let mir = Mir::utf8("[a-z]*").unwrap();
 
-        assert_eq!(len, 0);
+        assert_eq!(mir.priority(), 0);
+
+        let leaf = graph.push(Node::Leaf("LEAF"));
+        let id = graph.regex(mir, leaf);
+
         assert_eq!(
-            graph[parsed],
+            graph[id],
             Node::Fork(
                 Fork::new()
-                    .branch('a'..='z', parsed) // goto self == loop
+                    .branch('a'..='z', id) // goto self == loop
                     .miss(leaf)
             ),
         );
@@ -398,38 +301,20 @@ mod tests {
     fn maybe() {
         let mut graph = Graph::new();
 
-        let leaf = graph.push(Node::Leaf("LEAF"));
-        let (len, parsed) = graph.regex(true, "[a-z]?", leaf).unwrap();
+        let mir = Mir::utf8("[a-z]?").unwrap();
 
-        assert_eq!(len, 0);
+        assert_eq!(mir.priority(), 0);
+
+        let leaf = graph.push(Node::Leaf("LEAF"));
+        let id = graph.regex(mir, leaf);
+
         assert_eq!(
-            graph[parsed],
+            graph[id],
             Node::Fork(
                 Fork::new()
                     .branch('a'..='z', leaf)
                     .miss(leaf)
             ),
         );
-    }
-
-    #[test]
-    fn priorities() {
-        let mut graph = Graph::new();
-
-        let leaf = graph.push(Node::Leaf("LEAF"));
-
-        let regexes = [
-            ("[a-z]+", 1),
-            ("a|b", 2),
-            ("a|[b-z]", 1),
-            ("(foo)+", 6),
-            ("foobar", 12),
-            ("(fooz|bar)+qux", 12),
-        ];
-
-        for (regex, expected) in regexes.iter() {
-            let (len, _) = graph.regex(true, regex, leaf).unwrap();
-            assert_eq!(len, *expected);
-        }
     }
 }
